+++ pkg/apis/scheduling/v1alpha2/bindrequest_types.go
@@ -37,7 +37,7 @@ type ReceivedGPU struct {
 	// Count is the amount of GPUs devices that were received
 	Count int `json:"count,omitempty"`
 
-	// This is the portion size that the pod will receive from each connected gpu device
+	// This is the portion size that the pod will receive from each connected GPU device
 	// This is a serialized float that should be written as a decimal point number.
 	Portion string `json:"portion,omitempty"`
 }